http uses \r\n instead of plain \n as line separators
and unlike most software on the internet, async_h1 is not tolerant to
this and crashes on the response. fix vsock-demo.py to return a proper
http response for now. but async_h1 should really do better here (and
not crash the whole process).

print the request for debugging too